solucoes/luisvenezian/02-telegram-bot/requirements.txt
python-telegram-bot==13.5
requests==2.27.1